pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uPrintGen"
version = "1.0.0"
description = "Microprint generator"

readme = "README.md"
authors = [{ name = "AlphaSteam", email = "sebastian.alfaro@outlook.cl" }]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["microprint", "generator"]

dependencies = [
    "svgwrite",
    "pillow",
]
requires-python = ">=3.0"

[project.optional-dependencies]

[project.urls]
Homepage = "https://github.com/AlphaSteam/uPrintGen"

[project.scripts]
generate_microprint = "uPrintGen.command_line:main"